+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # Rossynt Changelog
 
 ## [Unreleased]
+### Added
+- Support IntelliJ Platform 2022.1 - 93661519db241a33cefa2a72b18c63dfcc7c439f
+
 ## [213.12.1] - 2022-03-18
 ### Fixed
 - Fix #86, caused by threading issues - 0a406c6c31e02056a7a0501202f80b0912c05c31